+++ packages/tiny-cli/src/index.ts
@@ -2,4 +2,4 @@ export * from "./typed";
 export * from "./sub-command";
 export * from "./presets";
 export * from "./zod";
-export { ParseError, consoleErrorExtra } from "./utils";
+export { ParseError } from "./utils";
