+++ packages/utils/src/index.ts
@@ -8,3 +8,4 @@ export * from "./promise";
 export * from "./misc";
 export * from "./hash";
 export * from "./regexp";
+export * from "./error";
