+++ packages/utils/src/misc.ts
@@ -7,21 +7,6 @@ export const typedBoolean = Boolean as unknown as <T>(
   value: T
 ) => value is Exclude<T, false | 0 | "" | null | undefined>;
 
-// traverse Error.cause chain
-export function flattenErrorCauses(e: unknown): unknown[] {
-  let errors: unknown[] = [e];
-  for (let i = 0; ; i++) {
-    if (i > 100) throw new Error("bound loop just in case");
-    if (e instanceof Error && e.cause && !errors.includes(e.cause)) {
-      errors.push(e.cause);
-      e = e.cause;
-      continue;
-    }
-    break;
-  }
-  return errors;
-}
-
 // similar convenience as z.enum but supports any key (number | string | symbol)
 // cf. https://github.com/colinhacks/zod/blob/cfbc7b3f6714ced250dd4053822faf472bf1828e/src/types.ts#L3958
 export function arrayToEnum<
